+++ CHANGELOG.md
@@ -959,6 +959,7 @@
 - [Support runtime checks of intersection types][7769]
 - [Merge `Small_Integer` and `Big_Integer` types][7636]
 - [Inline type ascriptions][7796]
+- [New `project/status` route for reporting LS state][7801]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -1100,6 +1101,7 @@
 [7769]: https://github.com/enso-org/enso/pull/7769
 [7636]: https://github.com/enso-org/enso/pull/7636
 [7796]: https://github.com/enso-org/enso/pull/7796
+[7801]: https://github.com/enso-org/enso/pull/7801
 
 # Enso 2.0.0-alpha.18 (2021-10-12)